#7 utility関連のソース整理
